+++ manyworlds/__init__.py
@@ -1,4 +1,4 @@
 from .feature import Feature
 
 __all__ = ["Feature"]
-__version__ = "0.4.1"
+__version__ = "0.5.0"