+++ manyworlds/feature.py
@@ -21,11 +21,36 @@ class Feature:
     The number of spaces per indentation level
     """
 
-    FEATURE_PATTERN: re.Pattern = re.compile("^Feature: (?P<feature_name>.*)")
+    FEATURE_PATTERN: re.Pattern = re.compile(
+        r"""
+        ^                    # start of line
+        Feature:             # "Feature:" keyword
+        [ ]                  # space
+        (?P<feature_name>.*) # feature name
+        $                    # end of line
+        """,
+        re.VERBOSE,
+    )
     """
     re.Pattern
 
-    The string "Feature: ", followed by arbitrary string
+    Pattern describing a BDD feature line ("Feature: …")
+    """
+
+    COMMENT_PATTERN: re.Pattern = re.compile(
+        r"""
+        ^                    # start of line
+        \#                   # "#" character
+        [ ]                  # space
+        (?P<comment>.*)      # comment
+        $                    # end of line
+        """,
+        re.VERBOSE,
+    )
+    """
+    re.Pattern
+
+    Pattern describing a comment line ("# …")
     """
 
     graph: ig.Graph
@@ -155,6 +180,7 @@ class Feature:
                 if scenario_match is not None:
                     feature.append_scenario(
                         scenario_match.group("scenario_name"),
+                        comment=scenario_match.group("comment"),
                         at_level=level,
                         line_no=line_no,
                     )
@@ -174,6 +200,11 @@ class Feature:
                     )
                     continue
 
+                # Comment line?
+                comment_match: Optional[re.Match] = cls.COMMENT_PATTERN.match(line)
+                if comment_match is not None:
+                    continue  # skip comment lines
+
                 # Feature description line?
                 if feature.name is not None and len(feature.scenarios()) == 0:
                     feature.description.append(line)
@@ -189,7 +220,7 @@ class Feature:
         return feature
 
     def append_scenario(
-        self, scenario_name: str, at_level: int, line_no: int
+        self, scenario_name: str, comment: Optional[str], at_level: int, line_no: int
     ) -> Scenario:
         """Append a scenario to the feature.
 
@@ -197,6 +228,9 @@ class Feature:
         ----------
         scenario : Scenario
             The scenario to append
+
+        comment : str, optional
+            A comment
 
         at_level : int
             The indentation level of the scenario in the input file.
@@ -220,6 +254,7 @@ class Feature:
                     scenario_name,
                     self.graph,
                     parent_scenario=parent_level_scenarios[-1],
+                    comment=comment,
                 )
             else:
                 raise InvalidFeatureFileError(
@@ -229,7 +264,7 @@ class Feature:
                 )
 
         else:  # Root scenario:
-            return Scenario(scenario_name, self.graph)
+            return Scenario(scenario_name, self.graph, comment=comment)
 
     def append_step(self, step: Step, at_level: int, line_no: int) -> None:
         """Appends a step to the feature.
@@ -311,7 +346,7 @@ class Feature:
 
     @classmethod
     def write_scenario_name(
-        cls, file_handle: TextIO, scenarios: List[Scenario]
+        cls, file_handle: TextIO, scenarios: List[Scenario], write_comment: bool = False
     ) -> None:
         """Writes formatted scenario name to the end of a "relaxed" flat feature file.
 
@@ -322,6 +357,9 @@ class Feature:
 
         scenarios : List[Scenario]
             Organizational and validated scenarios along the path
+
+        write_comment : bool, default = False
+            Whether or not to write comment if present
         """
 
         # (1) Group consecutive regular or organizational scenarios:
@@ -354,14 +392,22 @@ class Feature:
             else:
                 group_strings.append(" > ".join([sc.name for sc in group]))
 
-        # (3) Assemble and write name:
-        file_handle.write(
-            "Scenario: {scenario_name}\n".format(scenario_name=" ".join(group_strings))
-        )
+        # (3) Assemble name:
+        scenario_string: str = "Scenario: {}".format(" ".join(group_strings))
+
+        # (4) Optional comment:
+        destination_scenario: Scenario = scenarios[-1]
+        if write_comment is True and destination_scenario.comment is not None:
+            scenario_string += " # {comment}".format(
+                comment=destination_scenario.comment
+            )
+
+        # (4) Write name:
+        file_handle.write(scenario_string + "\n")
 
     @classmethod
     def write_scenario_steps(
-        cls, file_handle: TextIO, steps: List[Step], comments: bool = False
+        cls, file_handle: TextIO, steps: List[Step], write_comments: bool = False
     ) -> None:
         """Writes formatted scenario steps to the end of the flat feature file.
 
@@ -373,25 +419,29 @@ class Feature:
         steps : List[Step]
             Steps to append to file_handle
 
-        comments: bool
-            Whether or not to write comments
+        write_comments: bool, default = False
+            Whether or not to write comments if present
         """
 
         last_step: Optional[Step] = None
         for step in steps:
-            first_of_type = (
+            first_of_type: bool = (
                 last_step is None or last_step.conjunction != step.conjunction
             )
-            file_handle.write(step.format(first_of_type=first_of_type) + "\n")
-            if comments and step.comment:
-                file_handle.write("# {comment}\n".format(comment=step.comment))
+            step_string: str = step.format(first_of_type=first_of_type)
+            if write_comments is True and step.comment is not None:
+                step_string += " # {comment}".format(comment=step.comment)
+            file_handle.write(step_string + "\n")
+
             if step.data:
-                Feature.write_data_table(file_handle, step.data, comments)
+                Feature.write_data_table(
+                    file_handle, step.data, write_comment=write_comments
+                )
             last_step = step
 
     @classmethod
     def write_data_table(
-        cls, file_handle: TextIO, data_table: DataTable, comments: bool = False
+        cls, file_handle: TextIO, data_table: DataTable, write_comment: bool = False
     ) -> None:
         """Writes formatted data table to the end of the flat feature file.
 
@@ -403,8 +453,8 @@ class Feature:
         data_table : DataTable
             A data table
 
-        comments : bool
-            Whether or not to write comments
+        write_comment : bool
+            Whether or not to write comment if present
         """
 
         # Determine column widths to accommodate all values:
@@ -426,7 +476,7 @@ class Feature:
             )
 
             # add comments:
-            if comments and row.comment:
+            if write_comment is True and row.comment is not None:
                 table_row_string += " # {comment}".format(comment=row.comment)
 
             # write line:
@@ -436,7 +486,7 @@ class Feature:
         self,
         file_path: str,
         mode: Literal["strict", "relaxed"] = "strict",
-        comments: bool = False,
+        write_comments: bool = False,
     ) -> None:
         """Writes a flat (no indentation) feature file representing the feature.
 
@@ -459,11 +509,11 @@ class Feature:
 
             # Scenarios:
             if mode == "strict":
-                self.flatten_strict(flat_file, comments=comments)
+                self.flatten_strict(flat_file, write_comments=write_comments)
             elif mode == "relaxed":
-                self.flatten_relaxed(flat_file, comments=comments)
+                self.flatten_relaxed(flat_file, write_comments=write_comments)
 
-    def flatten_strict(self, flat_file: TextIO, comments: bool = False) -> None:
+    def flatten_strict(self, flat_file: TextIO, write_comments: bool = False) -> None:
         """Write. a flat (no indentation) feature file representing the feature
         using the "strict" flattening mode.
 
@@ -476,7 +526,7 @@ class Feature:
         flat_file : io.TextIOWrapper
             The flat feature file
 
-        comments : bool, default = False
+        write_comments : bool, default = False
             Whether or not to write comments
         """
 
@@ -487,7 +537,9 @@ class Feature:
                 for sc in scenario.path_scenarios()
                 if sc.is_organizational() or sc == scenario
             ]
-            Feature.write_scenario_name(flat_file, scenarios_for_naming)
+            Feature.write_scenario_name(
+                flat_file, scenarios_for_naming, write_comment=write_comments
+            )
 
             ancestor_scenarios = scenario.ancestors()
             steps: List[Step] = []
@@ -499,10 +551,12 @@ class Feature:
             steps += scenario.steps
 
             # Write steps:
-            Feature.write_scenario_steps(flat_file, steps, comments=comments)
+            Feature.write_scenario_steps(
+                flat_file, steps, write_comments=write_comments
+            )
             flat_file.write("\n")  # Empty line to separate scenarios
 
-    def flatten_relaxed(self, flat_file: TextIO, comments: bool = False) -> None:
+    def flatten_relaxed(self, flat_file: TextIO, write_comments: bool = False) -> None:
         """Writes a flat (no indentation) feature file representing the feature
         using the "relaxed" flattening mode.
 
@@ -515,8 +569,8 @@ class Feature:
         flat_file : io.TextIOWrapper
             The flat feature file
 
-        comments : bool, default = False
-            Whether or not to write comments
+        write_comments : bool, default = False
+            Whether or not to write comments if present
         """
 
         for scenario in self.leaf_scenarios():
@@ -533,9 +587,13 @@ class Feature:
                     path_scenario.validated = True
                     scenarios_for_naming.append(path_scenario)
 
-            Feature.write_scenario_name(flat_file, scenarios_for_naming)
+            Feature.write_scenario_name(
+                flat_file, scenarios_for_naming, write_comment=write_comments
+            )
             # Write steps:
-            Feature.write_scenario_steps(flat_file, steps, comments=comments)
+            Feature.write_scenario_steps(
+                flat_file, steps, write_comments=write_comments
+            )
             flat_file.write("\n")  # Empty line to separate scenarios
 
     def find(self, *scenario_names: List[str]) -> Optional[Scenario]: